pyproject.toml
# test it works
#   pip install -e .
# build the package
#   pip install build
#   python -m build
# check with twine
#   pip install twine
#   twine check dist/*

# push on testpypi
#   twine upload --repository testpypi dist/*
# install from testpypi
#   pip install -i https://test.pypi.org/simple/ silly_voice_lab

# push on pypi
#   twine upload dist/*



[build-system]
requires = ["wheel", "setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "silly-voice-lab"
version = "0.2.5"
description = "Dialogues voice generator using ElevenLabs (AI)"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
  {name = "Vincent Fabre", email = "peigne.plume@gmail.com"},
]
keywords= ["silly-voice-lab", "voice", "elevenlabs"]

classifiers=[
    "Programming Language :: Python :: 3",
    "Development Status :: 5 - Production/Stable",
    "Natural Language :: English",
]
dependencies = [
    "pydub>=0.25.1,<1.0.0",
    "pyttsx3>=2.99,<3.0.0",
    "PyYAML>=6.0.2,<7.0.0",
    "requests>=2.32.4,<3.0.0",
]

[project.scripts]
silly-voice-lab = "silly_voice_lab.main:cmd"

[project.urls]
Homepage = "https://github.com/byoso/ElevenLabs-tools"
Repository = "https://github.com/byoso/ElevenLabs-tools"

[tool.setuptools]
packages = [
    "silly_voice_lab",
    "silly_voice_lab.src",
    "silly_voice_lab.src.silly_engine",
    "silly_voice_lab.src.init_project",
    "silly_voice_lab.src.init_project.init",
    "silly_voice_lab.src.init_project.expert",
    "silly_voice_lab.src.init_project.init.scenario",
    "silly_voice_lab.src.init_project.expert.scenario",
    "silly_voice_lab.src.init_project.expert.configs",
    "silly_voice_lab.src.init_project.init.scenario.heroes",
    "silly_voice_lab.src.init_project.init.scenario.heroes",
    "silly_voice_lab.src.init_project.init.scenario.heroes.Sarah",
    "silly_voice_lab.src.init_project.init.scenario.heroes.John",
    ]

[tool.setuptools.package-data]
"*" = ["*.cfg", "*.yaml", "Makefile"]
